remove sliding connections
