+++ BrightID/src/components/Connections/ConnectionCard.js
@@ -8,9 +8,11 @@ import { useFocusEffect, useNavigation } from '@react-navigation/native';
 import moment from 'moment';
 import { CHANNEL_TTL } from '@/utils/constants';
 import { photoDirectory } from '@/utils/filesystem';
-import { staleConnection } from '@/actions';
+import { staleConnection, deleteConnection } from '@/actions';
 import verificationSticker from '@/static/verification-sticker.svg';
-import { DEVICE_LARGE } from '@/utils/deviceConstants';
+import { DEVICE_LARGE, WIDTH } from '@/utils/deviceConstants';
+import Material from 'react-native-vector-icons/MaterialCommunityIcons';
+import { useActionSheet } from '@expo/react-native-action-sheet';
 
 /**
  * Connection Card in the Connections Screen
@@ -134,6 +136,36 @@ const ConnectionCard = (props) => {
     }
   };
 
+  const showRemove = status === 'deleted' || status === 'stale';
+
+  const { showActionSheetWithOptions } = useActionSheet();
+  const removeOptions = ['Remove', 'cancel'];
+
+  const RemoveConnection = () =>
+    showRemove ? (
+      <TouchableOpacity
+        style={styles.removeButton}
+        onPress={() => {
+          showActionSheetWithOptions(
+            {
+              options: removeOptions,
+              cancelButtonIndex: removeOptions.length - 1,
+              destructiveButtonIndex: 0,
+              title: `Remove connection`,
+              message: `Are you sure you want to remove connection with ${name}? You can reconnect anytime.`,
+            },
+            (index) => {
+              if (index === 0) dispatch(deleteConnection(id));
+            },
+          );
+        }}
+      >
+        <Material color="#333" name="close" size={DEVICE_LARGE ? 22 : 18} />
+      </TouchableOpacity>
+    ) : (
+      <View />
+    );
+
   const imageSource =
     photo?.filename && !imgErr
       ? {
@@ -168,10 +200,10 @@ const ConnectionCard = (props) => {
           }}
           accessibilityLabel="View Connection details"
         >
-          <View style={styles.info}>
-            <View style={styles.nameContainer}>
+          <View style={[styles.info, { maxWidth: WIDTH * 0.56 }]}>
+            <View style={[styles.nameContainer]}>
               <Text
-                adjustsFontSizeToFit={true}
+                // adjustsFontSizeToFit={true}
                 numberOfLines={1}
                 style={styles.name}
                 testID="connectionCardText"
@@ -190,6 +222,7 @@ const ConnectionCard = (props) => {
             <ConnectionStatus />
           </View>
         </TouchableOpacity>
+        <RemoveConnection />
       </View>
     </View>
   );
@@ -258,7 +291,7 @@ const styles = StyleSheet.create({
   },
   waitingMessage: {
     fontFamily: 'Poppins-Medium',
-    fontSize: DEVICE_LARGE ? 14 : 12,
+    fontSize: DEVICE_LARGE ? 13 : 11,
     color: '#e39f2f',
     marginTop: DEVICE_LARGE ? 2 : 0,
   },
@@ -269,9 +302,13 @@ const styles = StyleSheet.create({
     marginTop: DEVICE_LARGE ? 5 : 2,
     textTransform: 'capitalize',
   },
-
   verificationSticker: {
     marginLeft: DEVICE_LARGE ? 5 : 3.5,
+  },
+  removeButton: {
+    width: DEVICE_LARGE ? 36 : 32,
+    position: 'absolute',
+    right: 0,
   },
 });
